Updated TinyDBDatabase.
Normalize all inputs parameter to receive and output
a `DictItem` (a normal dictionary with strings for keys).
This is important  because the conversion between simple dict and Item models will
be exclusive done in the `Model` class. All `Database`s should work plain dictionaries.

Moved `get_database` inside as a static method. This lets separate the creation of the database
for testing (in memory, for example) to the production database (which must be in disk).