Update Model.
Model is the main class for this library.
You can now inject a custom__item_schema to dynamically set the Pydantic model
used inside this class. This allow us to change the model outside the application if necessary.

Normalized all inputs to be DictItem

TODO: Normalize all inputs to be OutItem