Added functions to dynamically create the schema.
A needed feature is to be able to configure the database row schema outside
the library. These functions are the most important ones for this.
`augment_schema` reads a set-up yaml file and adds fields dynamically to
a pydantic model.
`load_schema_py` loads a whole pydantic model from a python file and uses
it as the model instead of the default.
TODO: Add validation for the custom model